+++ src/aks-preview/azext_aks_preview/tests/latest/test_helpers.py
@@ -4,7 +4,7 @@
 # --------------------------------------------------------------------------------------------
 
 import unittest
-from azure.cli.command_modules.acs import _helpers as helpers
+import azext_aks_preview._helpers as helpers
 
 
 class TestPopulateApiServerAccessProfile(unittest.TestCase):
@@ -51,6 +51,13 @@ class TestTrimContainerName(unittest.TestCase):
         container_name = 'abcdef-dns-ed55ba6d-hcp-centralus-azmk8s-io'
         expected_container_name = 'abcdef-dns-ed55ba6d'
         trim_container_name = helpers._trim_fqdn_name_containing_hcp(container_name)
+        self.assertEqual(expected_container_name, trim_container_name)
+
+    def test_trim_fqdn_name_trailing_dash(self):
+        container_name = 'dns-ed55ba6ad-e48fe2bd-b4bc-4aac-bc23-29bc44154fe1-privatelink-centralus-azmk8s-io'
+        expected_container_name = 'dns-ed55ba6ad-e48fe2bd-b4bc-4aac-bc23-29bc44154fe1-privatelink'
+        trim_container_name = helpers._trim_fqdn_name_containing_hcp(
+            container_name)
         self.assertEqual(expected_container_name, trim_container_name)
 
     def test_trim_fqdn_name_not_containing_hcp(self):